\name{riley}
\alias{riley}
\docType{data}
\title{
Riley Cambrian trilobite dataset
}
\description{
A classic dataset for biostratigraphy from Palmer (1954). Cambrian trilobites from the Riley formation (7 sections, 62 taxa) augmented with a small number of hypothetical mid-range events and unpaired range-end events. Data set included in CONOP9 CD.
}
\usage{data(riley)}
\details{
An object of class `CONOP`.
}
\source{
Palmer, 1954. The faunas of the Riley Formation in Central Texas. Journal of Paleontology, 28: 709-786.
}
\references{
Sadler, 2007. CONOP version 7.43.
}
\examples{
data(riley)
}
\keyword{datasets}
